requirements.txt
python-dotenv
PyMuPDF
openai
slack-sdk
langchain
PyPDF2
faiss-cpu
tiktoken
langchain_community
langchain_openai